PycharmProjects/Recommendation System_/requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
et-xmlfile==1.1.0
fonttools==4.43.1
forex-python==1.8
gensim==4.3.2
idna==3.4
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.1
nltk==3.8.1
numpy==1.26.1
openpyxl==3.2.0b1
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
punktdict==0.11
pyparsing==3.1.1
python-crfsuite==0.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
pyvi==0.1.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.3
simplejson==3.19.2
six==1.16.0
sklearn-crfsuite==0.3.6
smart-open==6.4.0
tabulate==0.9.0
threadpoolctl==3.2.0
tqdm==4.66.1
tzdata==2023.3
underthesea==6.8.0
underthesea_core==1.0.4
urllib3==2.0.7
